Separate mathjaxify function into a separate module
